Add progress indicators to cheat sheet
[why]
The progress indicators can not be found, people need to know that they
are there.

The progress indicators (taken from Fira Code) were introduced with #1733
and then moved:
    bd5936856  font-patcher: Add progress indicators
    41f91bcb4  Combine extraglyphs fonts

But no entry in the cheat sheet has been added. Which is the same for
the boxdrawing glyphs and others in extraglyphs, but those other will
probably be not searched for.

[how]
Add a new i_*.sh file for the extraglyphs icon set and introduce a new
prefix just for the progress indicators to improve discoverability.

Fixes: #1827

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>